refactor: flash mutation 파일 분리
