+++ src/api/flash/hook.ts
@@ -1,8 +1,6 @@
-import { getFlash, getFlashList, postFlash, putFlash } from '@api/flash';
+import { getFlash, getFlashList } from '@api/flash';
 import FlashQueryKey from '@api/flash/FlashQueryKey';
-import { serializeFlashFormData } from '@api/flash/serialize';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { FlashFormType } from '@type/form';
+import { useQuery } from '@tanstack/react-query';
 
 export const useFlashQuery = ({ meetingId }: { meetingId: number }) => {
   return useQuery({
@@ -15,26 +13,5 @@ export const useFlashListQuery = () => {
   return useQuery({
     queryKey: FlashQueryKey.list(),
     queryFn: getFlashList,
-  });
-};
-
-export const usePutFlashMutation = ({ meetingId }: { meetingId: number }) => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: ({ id, formData }: { id: number; formData: FlashFormType }) =>
-      putFlash(id, serializeFlashFormData(formData)),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: FlashQueryKey.detail(meetingId) });
-    },
-  });
-};
-
-export const usePostFlashMutation = () => {
-  return useMutation({
-    mutationFn: (formData: FlashFormType) => postFlash(serializeFlashFormData(formData)),
-    onError: () => {
-      alert('번쩍을 개설하지 못했습니다.');
-    },
   });
 };
